+++ package.json
@@ -16,8 +16,10 @@
     "@react-navigation/bottom-tabs": "^6.5.2",
     "@react-navigation/native": "^6.1.1",
     "@react-navigation/native-stack": "^6.9.7",
+    "date-fns": "^2.29.3",
     "react": "18.1.0",
     "react-native": "0.70.6",
+    "react-native-calendars": "^1.1293.0",
     "react-native-gesture-handler": "^2.9.0",
     "react-native-linear-gradient": "^2.6.2",
     "react-native-reanimated": "^2.14.1",
