--- src/components/calendarDrawer/customBackdrop/styles.ts
@@ -1,5 +1,0 @@
-import { StyleSheet, Platform } from "react-native";
-import { Fonts } from "@styles/font";
-import { EBrandColor, EFontColor } from "@styles/color";
-
-export const styles = StyleSheet.create({});
